requirements.txt
flake8==4.0.1
Flask==1.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.38.2
python-dotenv==0.19.2
itsdangerous==1.1.0